Control spec file order with an explicit list
The previous method of ordering specification chapters relied on numeric
prefixes in the filenames. This was brittle and made reordering or inserting
new sections difficult, as it required renaming multiple files.

This change removes the numeric prefixes from the markdown filenames. A new
`spec_files.txt` file is introduced to explicitly define the order of the
files for document generation. The Makefile is updated to use this file list
instead of a glob pattern, making the build process more robust and easier to
maintain.

Signed-off-by: Simo Sorce <simo@redhat.com>