Update spec for CKM_PUB_KEY_FROM_PRIV_KEY
Rewrite the documentation for deriving a public key from a private key to
improve clarity and correctness.

The list of default attributes for the new public key has been reformatted
into a table. The text now explicitly states that this mechanism can be used
regardless of the private key's CKA_DERIVE attribute.

The list of handled attributes has also been updated, adding CKA_ENCAPSULATE
and removing several attributes not applicable to public keys.

Signed-off-by: Simo Sorce <simo@redhat.com>